fixed lst_set_int in same way as lst_push_int (Brian Ripley fix)
git-svn-id: svn+ssh://nextgen.cshl.edu/usr/local/svnroot/phast/trunk@6468 3e584e52-61a1-4911-b546-322c73959ec4

+++ src/lib/base/list_of_lists.c
@@ -38,7 +38,7 @@ void lol_push(ListOfLists *lol, void *data, const char *name,
     strcpy(tempname, name);
   }
   lst_push_ptr(lol->lstName, (void*)tempname);
-  lst_push_int(lol->lstType, (int)listType);
+  lst_push_int(lol->lstType, listType);
 }
